Create utils/relationship
